install sanctum

+++ database/migrations/2021_07_21_174300_create_books_table.php
@@ -15,7 +15,7 @@ class CreateBooksTable extends Migration
     {
         Schema::create('books', function (Blueprint $table) {
             $table->id();
-            $table->integer('isbn13');
+            $table->bigInteger('isbn13');
             $table->timestamps();
         });
     }
